refactor(Pawn): 흰색 폰과 검은색 폰 클래스 분리
